Removes AUTH_UPDATED_SESSION_REF config.
It is no longer needed and and causes unnecessary configuration complexity and
problems.

The AUTH_UPDATED_SESSION_REF was only in place to ease the transition from
placing the auth object on `request.session` to `request.auth`. Placing in at
`request.session` was a mistake during the NG because it interfered with flask
sessions.

Also bumps version from 1.0 to 1.1.